requirements_export.txt
# Dependências para o script de exportação CSV
pandas>=2.0.0
pymongo>=4.0.0
python-dotenv>=1.0.0